Extend waitForPage for customer app
